1.2.1: Authentication on relevant operations
- Added the authentication phase to the operations that needed it and documented why. The ones who don't have are also documented with reason why they don't have it.
- Added a function to UserAuth that allows to delete an entry from the cache to make the cookie invalid. I don't know if we need to also delete the cookie from the request.
- Added back the documentation that was deleted in a previous commit.
- Memory side also has the cookie though not used entirely.

This was not tested yet so someone else could please run the tests to see if they mess up. The ones we already have should be sufficient to check.

+++ src/scc/azure/AzureMonolithService.java
@@ -9,6 +9,7 @@ import java.util.stream.Collectors;
 import com.azure.cosmos.CosmosClientBuilder;
 import com.azure.cosmos.models.CosmosPatchOperations;
 
+import jakarta.ws.rs.core.Cookie;
 import scc.azure.config.AzureMonolithConfig;
 import scc.azure.dao.AuctionDAO;
 import scc.azure.dao.BidDAO;
@@ -56,17 +57,24 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
     }
 
     /**
-     * Creates an auction
-     * 
+     * Creates an auction for a user with given identifier.
+     * The user creating the auction must be logged in to execute this operation.
      * @param params JSON that contains the necessary information to create an
      *               auction
+     * @param auth Cookie related to the user being "logged" in the application
      * @return 200 with auction's identifier if successful,
+     * 404 if the user does not exist
+     * 403 if the authentication phase failed
      */
     @Override
-    public Result<String, ServiceError> createAuction(CreateAuctionParams params) {
+    public Result<String, ServiceError> createAuction(CreateAuctionParams params, Cookie auth) {
         var validateResult = AuctionService.validateCreateAuctionParams(params);
         if (validateResult.isError())
             return Result.err(validateResult.error());
+
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
 
         if (!this.userDB.userExists(params.userId()))
             return Result.err(ServiceError.USER_NOT_FOUND);
@@ -84,19 +92,48 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return response.map(AuctionDAO::getId);
     }
 
+    /**
+     * Sets the auction state into deleted so, when displayed to the user, it will appear "created by DELETED USER".
+     * This operation does not need authentication as the DeleteUser already did it.
+     * @param auctionId identifier of the auction to be deleted
+     * @return 204 if deleted successfully,
+     * 404 if the auction doesn't exist (do we really need to check this?)
+     */
     @Override
     public Result<Void, ServiceError> deleteAuction(String auctionId) {
         return this.auctionDB.deleteAuction(auctionId);
     }
 
+    /**
+     * Lists all auctions created by a user with given identifier.
+     * This operation does not need authentication as
+     * any user should be able to see the auctions from any other user, even their own.
+     * @param userId identifier of the user
+     * @return 200 with all user's auctions or empty,
+     * 404 if the user does not exist
+     */
     @Override
     public Result<List<String>, ServiceError> listAuctionsOfUser(String userId) {
         return this.auctionDB.listAuctionsOfUser(userId)
                 .map(auctions -> auctions.stream().map(AuctionDAO::getId).collect(Collectors.toList()));
     }
 
+    /**
+     * Updates all the auction's values into new given ones.
+     * The user who is the owner of the auction must be logged in order to execute this operation.
+     * @param auctionId identifier of the auction to be updated
+     * @param ops Values that are to be changed to new ones
+     * @param auth Cookie related to the user being "logged" in the application
+     * @return 204 if successful on changing the auction's values,
+     * 404 if the auction does not exist,
+     * 403 if the authentication phase failed
+     */
     @Override
-    public Result<Void, ServiceError> updateAuction(String auctionId, UpdateAuctionOps ops) {
+    public Result<Void, ServiceError> updateAuction(String auctionId, UpdateAuctionOps ops, Cookie auth) {
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
+
         var patchOps = CosmosPatchOperations.create();
         if (ops.shouldUpdateTitle())
             patchOps.set("/title", ops.getTitle());
@@ -114,10 +151,24 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return result;
     }
 
+    /**
+     * Creates a bid in an auction with given identifier from the logged user. If the bid overpowers all current ones,
+     * the auction's winning bid will be set to this one's identifier
+     * The user making the bid must be logged in to execute this operation
+     * @param params Parameters required to create a bid object
+     * @param auth Cookie related to the user being "logged" in the application
+     * @return 200 with bid's generated identifier,
+     * 404 if the user making the bid or the auction doesn't exist
+     * 403 if the authentication phase failed
+     */
     @Override
-    public Result<String, ServiceError> createBid(CreateBidParams params) {
+    public Result<String, ServiceError> createBid(CreateBidParams params, Cookie auth) {
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
+
         if (!this.userDB.userExists(params.userId()))
-            return Result.err(ServiceError.AUCTION_NOT_FOUND);
+            return Result.err(ServiceError.USER_NOT_FOUND);
 
         if (!this.auctionDB.auctionExists(params.auctionId()))
             return Result.err(ServiceError.AUCTION_NOT_FOUND);
@@ -128,6 +179,14 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return response.map(BidDAO::getId);
     }
 
+    /**
+     * Lists all bids made on an auction with given identifier.
+     * This operation does not need authentication as any user should be
+     * able to see any auction's bids to know whether they bid on it or not.
+     * @param auctionId identifier of the auction
+     * @return 200 with list of bids in the auction or empty,
+     * 404 if auction doesn't exist
+     */
     @Override
     public Result<List<BidItem>, ServiceError> listBids(String auctionId) {
         if (!this.auctionDB.auctionExists(auctionId))
@@ -138,8 +197,21 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
                 .collect(Collectors.collectingAndThen(Collectors.toList(), Result::ok));
     }
 
+    /**
+     * Creates a question on an auction with given identifier made by a user with given identifier
+     * The user making the question must be logged in to make this operation
+     * @param params Parameters required to create a question
+     * @param auth Cookie related to the user being "logged" in the application
+     * @return 200 with generated question's identifier,
+     * 404 if the user or the auction does not exist,
+     * 403 if the authentication phase failed
+     */
     @Override
-    public Result<String, ServiceError> createQuestion(CreateQuestionParams params) {
+    public Result<String, ServiceError> createQuestion(CreateQuestionParams params, Cookie auth) {
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
+
         if (!this.userDB.userExists(params.userId()))
             return Result.err(ServiceError.USER_NOT_FOUND);
         if (!this.auctionDB.auctionExists(params.auctionId()))
@@ -151,8 +223,21 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return response.map(QuestionDAO::getId);
     }
 
+    /**
+     * Creates a reply destined towards a specific question made in a given auction that is answered by the owner
+     * The user who is owner of the auction must be logged in to execute this operation
+     * @param params Parameters required to create a reply
+     * @param auth Cookie related to the user being "logged" in the application
+     * @return 200 with generated reply's identifier,
+     * 404 if the user or auction does not exist,
+     * 403 if the authentication phase failed
+     */
     @Override
-    public Result<Void, ServiceError> createReply(CreateReplyParams params) {
+    public Result<Void, ServiceError> createReply(CreateReplyParams params, Cookie auth) {
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
+
         if (!this.userDB.userExists(params.userId()))
             return Result.err(ServiceError.USER_NOT_FOUND);
 
@@ -165,6 +250,12 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return Result.ok();
     }
 
+    /**
+     * Lists all the questions made in an auction with given identifier
+     * @param auctionId identifier of the auction
+     * @return 200 with the questions made in the auction or empty,
+     * 404 if the auction does not exist
+     */
     @Override
     public Result<List<QuestionItem>, ServiceError> listQuestions(String auctionId) {
         if (!this.auctionDB.auctionExists(auctionId))
@@ -178,26 +269,55 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
                 .collect(Collectors.collectingAndThen(Collectors.toList(), Result::ok));
     }
 
+    /**
+     * Uploads a media resource into the blob storage
+     * @param contents bytes of the media resource
+     * @return Uploaded media resource's generated identifier
+     */
     @Override
     public String uploadMedia(byte[] contents) {
         return this.mediaStorage.uploadAuctionMedia(contents);
     }
 
+    /**
+     * Uploads a media resource destined for a user with given identifier
+     * @param userId identifier of the user
+     * @param contents bytes of the media resource
+     * @return Uploaded media resource's generated identifier
+     */
     @Override
     public String uploadUserProfilePicture(String userId, byte[] contents) {
         return this.mediaStorage.uploadUserMedia(contents);
     }
 
+    /**
+     * Downloads the contents of a media resource with given identifier stored in the blob storage
+     * @param id identifier of the media resource
+     * @return bytes of the associated media resource
+     */
     @Override
     public Optional<byte[]> downloadMedia(String id) {
         return this.mediaStorage.downloadMedia(id);
     }
 
+    /**
+     * Deletes the contents of a media resource with given identifier from the blob storage
+     * @param id identifier of the media resource
+     * @return true if it was deleted, false otherwise
+     */
     @Override
     public boolean deleteMedia(String id) {
         return this.mediaStorage.deleteMedia(id);
     }
 
+    /**
+     * Creates a user with given values to be stored in the application's user database.
+     * This operation does not need authentication as anyone outside the application should be able to register
+     * to the application.
+     * @param params Parameters required to create a user object
+     * @return 200 with generated user's identifier,
+     * 409 if the user with given identifier already exists
+     */
     @Override
     public Result<String, ServiceError> createUser(CreateUserParams params) {
         var validateResult = UserService.validateCreateUserParams(params);
@@ -212,8 +332,23 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return result.map(UserDAO::getId);
     }
 
+    /**
+     * Deletes a user with given identifier from the application's user database.
+     * The user logged in must be the same as the one from the request and, after the operation is executed,
+     * we make the cookie invalid.
+     * All the auctions associated with this user must be in the state of deleted as in, when
+     * performing reads on any of them, we show that it was created by a "DELETED USER".
+     * @param userId identifier of the user
+     * @param auth Cookie related to the user being "logged" in the application
+     * @return 204 if deleted successfully,
+     * 403 if the authentication phase failed
+     */
     @Override
-    public Result<Void, ServiceError> deleteUser(String userId) {
+    public Result<Void, ServiceError> deleteUser(String userId, Cookie auth) {
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
+
         var result = this.userDB.deleteUser(userId);
         if (!result.isOk())
             return Result.err(result.error());
@@ -223,14 +358,33 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
             deleteMedia(photoId);
 
         auctionDB.deleteUserAuctions(userId);
+        //TODO: So which one do we do ??? Doing both is redundant as the first's effects are cancelled by second's
+
         // TODO set user in user bids as DELETED USER
         // TODO Delete the question/reply entries from this user
+
+        this.userAuth.deleteSessionToken(auth.getValue());
 
         return Result.ok();
     }
 
+    /**
+     * Updates the values stored in a user with given identifier to new ones
+     * The updated user must be logged in to execute this operation
+     * @param userId identifier of the user
+     * @param ops Values to be placed on the user
+     * @param auth Cookie related to the user being "logged" in the application
+     * @return 204 if updated successfully,
+     * 404 if the user does not exist,
+     * 403 if the authentication phase failed
+     * 400 if the request is not well-formed
+     */
     @Override
-    public Result<Void, ServiceError> updateUser(String userId, UpdateUserOps ops) {
+    public Result<Void, ServiceError> updateUser(String userId, UpdateUserOps ops, Cookie auth) {
+        var authResult = this.userAuth.validateSessionToken(auth.getValue());
+        if (authResult.isError())
+            return Result.err(authResult.error());
+
         var validateResult = UserService.validateUpdateUserOps(ops);
         if (validateResult.isError())
             return Result.err(validateResult.error());
@@ -252,6 +406,17 @@ public class AzureMonolithService implements UserService, MediaService, AuctionS
         return result;
     }
 
+    /**
+     * "Logs" in a user with given identifier and given password
+     * The user will be authenticated if the password stored in the application's user database
+     * matches the one given from the request and is given a sessionToken through a cookie with an expiration
+     * time of 30 minutes
+     * @param userId identifier of the user
+     * @param password password of the user
+     * @return 200 with generated session token,
+     * 404 if the user does not exist
+     * 403 if the passwords do not match
+     */
     @Override
     public Result<String, ServiceError> authenticateUser(String userId, String password) {
         return this.userAuth.authenticate(userId, password);
